drf-vue3-typescript-demo-blog/Blog-Backend/requirements.txt
django==2.2.23
djangorestframework==3.12.4
django-smtp-ssl==1.0
django-mptt==0.12.0
django-filter==2.4.0
drf-yasg==1.20.0